chore(readme): Fix incorrect 'endpoint' instructions in Exercise 2
This commit fixes an issue in the instructions for Exercise 2,
specifcally around the 'endpoint' variable assignment.

Previously, the instructions specified that the 'endpoint' variable should be
set equal to the 'comic' property of the 'ENDPOINTS' object. This has been
changed to the 'characater' property of the 'ENDPOINTS' object, which contains
the correct value.