chore(readme): Update SearchTypeControls example to match instructions
Previously, the 'SearchTypeControls' instructions in Exercise 4 included
a code snippet that, when copied directly into the application, caused
the application to crash. Additionally, there was a discrepnacy between
the code snippet and the instructions themselves, which described the
correct steps and approach.

This commit updates the code snippet such that copying it directly into
the application will result in the correct/expected behaviour.