chore(readme): Fix incorrect ResultsList example in Exercise 4
Previously, Exercise 4 included a code snippet that shows a prop of
'seachType' being passed down from the 'App' to the 'ResultsList'
component. This snippet also shows that the value of this prop is
exracted from 'this.prop.searchType', which suggests that the 'App'
itself receives a prop of 'searchType'. Since the 'App' does not receive
a 'searchType' prop, copying this snippet into the application results
in a bug.

This snippet has been updated to show that the value of the 'searchType'
prop is extracted from 'this.state.searchType'.